% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_dummy_data.R
\docType{data}
\name{binary_dummy_data}
\alias{binary_dummy_data}
\title{Binary dummy data}
\format{
A data.frame with 107 rows and 3 variables.
}
\description{
This data is made-up data for use in examples.
}
