% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_pse_auxiliary.R
\name{cal_delta_star}
\alias{cal_delta_star}
\title{Calculate delta star for sensitivity analysis}
\usage{
cal_delta_star(
  FR2max,
  R2,
  R2_uncond,
  est_eff,
  eff_thr,
  var_x,
  var_y,
  est_uncond,
  rxz,
  n_obs
)
}
\arguments{
\item{FR2max}{maximum R2}

\item{R2}{current R2}

\item{R2_uncond}{unconditional R2}

\item{est_eff}{estimated effect}

\item{eff_thr}{effect threshold}

\item{var_x}{variance of X}

\item{var_y}{variance of Y}

\item{est_uncond}{unconditional estimate}

\item{rxz}{correlation coefficient between X and Z}

\item{n_obs}{number of observations}
}
\value{
delta star value
}
\description{
Calculate delta star for sensitivity analysis
}
