% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_pse_auxiliary.R
\name{cal_rxy}
\alias{cal_rxy}
\title{Calculate rxy based on ryxGz, rxz, and ryz}
\usage{
cal_rxy(ryxGz, rxz, ryz)
}
\arguments{
\item{ryxGz}{correlation coefficient between Y and X given Z}

\item{rxz}{correlation coefficient between X and Z}

\item{ryz}{correlation coefficient between Y and Z}
}
\value{
rxy value
}
\description{
Calculate rxy based on ryxGz, rxz, and ryz
}
