% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_pse_auxiliary.R
\name{cal_rxz}
\alias{cal_rxz}
\title{Calculate R2xz based on variances and standard error}
\usage{
cal_rxz(var_x, var_y, R2, df, std_err)
}
\arguments{
\item{var_x}{variance of X}

\item{var_y}{variance of Y}

\item{R2}{coefficient of determination}

\item{df}{degrees of freedom}

\item{std_err}{standard error}
}
\value{
R2xz value
}
\description{
Calculate R2xz based on variances and standard error
}
