% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_pse_auxiliary.R
\name{cal_ryz}
\alias{cal_ryz}
\title{Calculate R2yz based on ryxGz and R2}
\usage{
cal_ryz(ryxGz, R2)
}
\arguments{
\item{ryxGz}{correlation coefficient between Y and X given Z}

\item{R2}{coefficient of determination}
}
\value{
R2yz value
}
\description{
Calculate R2yz based on ryxGz and R2
}
