% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_auxiliary.R
\name{chisq_p}
\alias{chisq_p}
\title{Perform a Chi-Square Test}
\usage{
chisq_p(a, b, c, d)
}
\arguments{
\item{a}{Frequency count for row 1, column 1.}

\item{b}{Frequency count for row 1, column 2.}

\item{c}{Frequency count for row 2, column 1.}

\item{d}{Frequency count for row 2, column 2.}
}
\value{
P-value from the chi-square test.
}
\description{
`chisq_p` calculates the p-value for a chi-square test given a contingency table.
}
