% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concord1.R
\docType{data}
\name{concord1}
\alias{concord1}
\title{Concord1 data}
\format{
A data.frame with 496 rows and 10 variables.
}
\usage{
data(concord1)
}
\description{
This data .. XXX
}
\details{
\tabular{lll}{
  \strong{sex} \tab \code{factor} \tab Self-identified sex of participants, Male, Female, or Transgendered.\cr
  \strong{age_years} \tab \code{numeric} \tab Participants' age in years.\cr
}
}
\references{
Curry, O. S., Jones Chesters, M., & Van Lissa, C. J. (2019). Mapping morality with a compass: Testing the theory of ‘morality-as-cooperation’ with a new questionnaire. Journal of Research in Personality, 78, 106–124.
\href{https://doi.org/10.1016/j.jrp.2018.10.008}{DOI:10.1016/j.jrp.2018.10.008}
}
