% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konfound-lmer.R
\name{get_kr_df}
\alias{get_kr_df}
\title{Extract Degrees of Freedom for Fixed Effects in a Linear Mixed-Effects Model}
\usage{
get_kr_df(model_object)
}
\arguments{
\item{model_object}{The mixed-effects model object produced by lme4::lmer.}
}
\value{
A vector containing degrees of freedom for the fixed effects in the model.
}
\description{
Extract Degrees of Freedom for Fixed Effects in a Linear Mixed-Effects Model
}
