% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konfound-glm.R
\name{konfound_glm}
\alias{konfound_glm}
\title{Konfound Analysis for Generalized Linear Models}
\usage{
konfound_glm(
  model_object,
  tested_variable_string,
  alpha,
  tails,
  index = "RIR",
  to_return
)
}
\arguments{
\item{model_object}{The model object produced by glm.}

\item{tested_variable_string}{The name of the variable being tested.}

\item{alpha}{Significance level for hypothesis testing.}

\item{tails}{Number of tails for the test (1 or 2).}

\item{index}{Type of sensitivity analysis ('RIR' by default).}

\item{to_return}{The type of output to return.}
}
\value{
The results of the konfound analysis for the specified variable(s).
}
\description{
This function performs konfound analysis on a generalized linear model
object. It uses 'broom' to tidy model outputs and calculates the sensitivity
of inferences. It supports analysis for a single variable
 or multiple variables.
}
