% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konfound-lm.R
\name{konfound_lm}
\alias{konfound_lm}
\title{Konfound Analysis for Linear Models}
\usage{
konfound_lm(
  model_object,
  tested_variable_string,
  alpha,
  tails,
  index,
  to_return
)
}
\arguments{
\item{model_object}{The linear model object produced by lm.}

\item{tested_variable_string}{The name of the variable being tested.}

\item{alpha}{Significance level for hypothesis testing.}

\item{tails}{Number of tails for the test (1 or 2).}

\item{index}{Type of sensitivity analysis ('RIR' by default).}

\item{to_return}{The type of output to return.}
}
\value{
The results of the konfound analysis for the specified variable(s).
}
\description{
This function performs konfound analysis on a linear model object 
produced by lm.
It calculates the sensitivity of inferences for coefficients in the model.
It supports analysis for a single variable or multiple variables.
}
