% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konfound-lmer.R
\name{konfound_lmer}
\alias{konfound_lmer}
\title{Konfound Analysis for Linear Mixed-Effects Models}
\usage{
konfound_lmer(
  model_object,
  tested_variable_string,
  test_all,
  alpha,
  tails,
  index,
  to_return
)
}
\arguments{
\item{model_object}{The mixed-effects model object produced by lme4::lmer.}

\item{tested_variable_string}{The name of the fixed effect being tested.}

\item{test_all}{Boolean indicating whether to test all fixed effects or not.}

\item{alpha}{Significance level for hypothesis testing.}

\item{tails}{Number of tails for the test (1 or 2).}

\item{index}{Type of sensitivity analysis ('RIR' by default).}

\item{to_return}{The type of output to return.}
}
\value{
The results of the konfound analysis for the specified fixed effect(s).
}
\description{
This function performs konfound analysis on a linear mixed-effects model
object produced by lme4::lmer. It calculates the sensitivity of inferences
for fixed effects in the model. It supports analysis for a single variable or multiple variables.
}
