% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{launch_shiny}
\alias{launch_shiny}
\title{Open codebook}
\usage{
launch_shiny()
}
\value{
Launches a web browser
}
\description{
Open codebook
}
\details{
Open the Shiny interactive web application in a browser
}
