% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkonfound.R
\name{mkonfound}
\alias{mkonfound}
\title{Perform meta-analyses including sensitivity analysis}
\usage{
mkonfound(d, t, df, alpha = 0.05, tails = 2, return_plot = FALSE)
}
\arguments{
\item{d}{data.frame or tibble with the t-statistics and associated degrees of freedom}

\item{t}{t-statistic or vector of t-statistics}

\item{df}{degrees of freedom or vector of degrees of freedom associated with the t-statistics in the t argument}

\item{return_plot}{whether to return a plot of the percent bias; defaults to FALSE}

\item{component_correlations}{whether to return the component correlations as part of the correlation-based approach}
}
\value{
prints the bias and the number of cases that would have to be replaced with cases for which there is no effect to invalidate the inference for each of the cases in the data.frame
}
\description{
For fitted models, this command carries out sensitivity analysis for a number of models, when their parameters stored in a data.frame.
}
\examples{
d <- read.csv("https://msu.edu/~kenfrank/example\%20dataset\%20for\%20mkonfound.csv")
str(d)
mkonfound(d, t, df)
}
