% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkonfound-data.R
\docType{data}
\name{mkonfound_ex}
\alias{mkonfound_ex}
\title{Example data for the mkonfound function}
\format{
A data frame with 30 rows and 2 variables:
\describe{
  \item{t}{t value}
  \item{df}{degrees of freedom associated with the t value}
  ...
}
}
\source{
\url{https://drive.google.com/file/d/1aGhxGjvMvEPVAgOA8rrxvA97uUO5TTMe/view}
}
\usage{
mkonfound_ex
}
\description{
A dataset containing t and df values from example studies from Educational Evaluation
and Policy Analysis (as detailed in Frank et al., 2013): 
https://drive.google.com/file/d/1aGhxGjvMvEPVAgOA8rrxvA97uUO5TTMe/view
}
\keyword{datasets}
