% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_output_dataframe.R
\name{output_df}
\alias{output_df}
\title{Output data frame based on model estimates and thresholds}
\usage{
output_df(
  est_eff,
  beta_threshhold,
  unstd_beta,
  bias = NULL,
  sustain = NULL,
  recase,
  obs_r,
  critical_r,
  r_con,
  itcv,
  non_linear
)
}
\arguments{
\item{est_eff}{estimated effect}

\item{beta_threshhold}{threshold for beta}

\item{unstd_beta}{unstandardized beta value}

\item{bias}{bias to change inference}

\item{sustain}{sustain to change inference}

\item{recase}{number of cases to replace null}

\item{obs_r}{observed correlation}

\item{critical_r}{critical correlation}

\item{r_con}{correlation for omitted variable}

\item{itcv}{inferential threshold for confounding variable}

\item{non_linear}{flag for non-linear models}
}
\value{
data frame with model information
}
\description{
Output data frame based on model estimates and thresholds
}
