% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_output_table.R
\name{output_table}
\alias{output_table}
\title{Output a Tidy Table from a Model Object}
\usage{
output_table(model_object, tested_variable)
}
\arguments{
\item{model_object}{A model object from which to generate the output.}

\item{tested_variable}{The variable being tested in the model.}
}
\value{
A tidy data frame containing model outputs, ITCV, 
and impacts for covariates.
}
\description{
This function takes a model object and the tested variable, 
tidies the model output using `broom::tidy`, 
calculates the impact threshold for confounding variables (ITCV) and impact 
for each covariate,and returns a rounded, tidy table of model outputs.
}
