% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plot_correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot Correlation Diagram}
\usage{
plot_correlation(r_con, obs_r, critical_r)
}
\arguments{
\item{r_con}{Correlation coefficient related to the confounding variable.}

\item{obs_r}{Observed correlation coefficient.}

\item{critical_r}{Critical correlation coefficient for decision-making.}
}
\value{
A ggplot object representing the correlation diagram.
}
\description{
This function creates a plot to illustrate the correlation between different 
variables,specifically focusing on the confounding variable, predictor of 
interest, and outcome.It uses ggplot2 for graphical representation.
}
