% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plot_empirical.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot the empirical k* distribution}
\usage{
plot_hist(result, type = c("stacked", "shape"), verbose = FALSE, ...)
}
\arguments{
\item{result}{A konfound_empdist object}

\item{type}{"stacked" (default, requires case_info) or "shape" (always available)}

\item{verbose}{If TRUE, prints a description of the plot to the console.
Default is FALSE.}

\item{...}{Passed to the underlying plot function}
}
\description{
Plot the empirical k* distribution
}
