% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plot_empirical.R
\name{plot_rir_dist}
\alias{plot_rir_dist}
\title{Distribution of per-case RIR influence}
\usage{
plot_rir_dist(result, verbose = FALSE, ...)
}
\arguments{
\item{result}{A konfound_empdist object (must have been run with case_info)}

\item{verbose}{If TRUE, prints a description of the plot to the console.
Default is FALSE.}

\item{...}{Passed to the underlying plot function}
}
\description{
Distribution of per-case RIR influence
}
