% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plot_threshold.R
\name{plot_threshold}
\alias{plot_threshold}
\title{Plot Effect Threshold Diagram}
\usage{
plot_threshold(beta_threshold, est_eff)
}
\arguments{
\item{beta_threshold}{The threshold value for the effect.}

\item{est_eff}{The estimated effect size.}
}
\value{
A ggplot object representing the effect threshold diagram.
}
\description{
This function creates a plot to illustrate the threshold of an effect 
estimate in relation to a specified beta threshold. It uses ggplot2 
for graphical representation.
}
