% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_sensitivity.R
\name{print.konfound}
\alias{print.konfound}
\title{Print method with concise summary of konfound output}
\usage{
\method{print}{konfound}(x, ...)
}
\arguments{
\item{...}{Additional arguments}

\item{object}{A `konfound` object}
}
\description{
Print method with concise summary of konfound output
}
\details{
Prints a concise summary of konfound output with multiple types of data specified in the to_return argument
}
