% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sensitivity.R
\name{summary.konfound}
\alias{summary.konfound}
\title{Concise summary of konfound output}
\usage{
\method{summary}{konfound}(object, ...)
}
\arguments{
\item{object}{A `konfound` object}

\item{...}{Additional arguments}
}
\description{
Concise summary of konfound output
}
\details{
Prints a concise summary of konfound output with multiple types of data specified in the to_return argument
}
