% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cop.R
\name{test_cop}
\alias{test_cop}
\title{Coefficient of Proportionality (COP) Test}
\usage{
test_cop(
  est_eff,
  std_err,
  n_obs,
  n_covariates,
  sdx,
  sdy,
  R2,
  eff_thr = 0,
  FR2max_multiplier = 1.3,
  FR2max = 0,
  alpha = 0.05,
  tails = 2,
  to_return = to_return
)
}
\arguments{
\item{est_eff}{The estimated effect (unstandardized).}

\item{std_err}{The standard error of the effect (unstandardized).}

\item{n_obs}{Number of observations.}

\item{n_covariates}{Number of covariates in the model.}

\item{sdx}{Standard deviation of the predictor variable.}

\item{sdy}{Standard deviation of the outcome variable.}

\item{R2}{R-squared of the model, not adjusted.}

\item{eff_thr}{Threshold for the effect size, unstandardized.}

\item{FR2max_multiplier}{Multiplier for R2 to get R2max.}

\item{FR2max}{Maximum R-squared in the final model with an 
unobserved confounder.}

\item{alpha}{Significance level for hypothesis testing (default: 0.05).}

\item{tails}{Number of tails for hypothesis testing (default: 2).}

\item{to_return}{Type of output to return ('raw_output', 'print', or other).}
}
\value{
A list containing results of the COP test, including delta star,
 delta exact,percentage bias, and other statistical measures.
  Can also print summary results.
}
\description{
Conducts the Coefficient of Proportionality (COP) test, calculating both
Oster's approximate and exact versions of COP.
}
