% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_sensitivity.R
\name{test_sensitivity}
\alias{test_sensitivity}
\title{Print method with concise summary of konfound output}
\usage{
test_sensitivity(unstd_beta, std_err, n_obs, n_covariates, alpha, tails, nu,
  to_return, component_correlations, non_linear, model_object, tested_variable,
  t, df)
}
\arguments{
\item{object}{A `konfound` object}

\item{...}{Additional arguments}
}
\description{
Print method with concise summary of konfound output
}
\details{
Prints a concise summary of konfound output with multiple types of data specified in the to_return argument
}
