% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sensitivity_ln.R
\name{test_sensitivity_ln}
\alias{test_sensitivity_ln}
\title{Test Sensitivity for Non-linear Models}
\usage{
test_sensitivity_ln(
  est_eff,
  std_err,
  n_obs,
  n_covariates,
  n_treat,
  switch_trm = T,
  replace = "control",
  alpha,
  tails,
  nu,
  to_return,
  model_object,
  tested_variable
)
}
\arguments{
\item{est_eff}{Estimated effect size.}

\item{std_err}{Standard error of the estimated effect.}

\item{n_obs}{Number of observations in the study.}

\item{n_covariates}{Number of covariates in the model.}

\item{n_treat}{Number of cases in the treatment group.}

\item{switch_trm}{Logical value indicating whether to switch 
treatment and control groups (default: TRUE).}

\item{replace}{Specifies the group for base rate calculation 
('control' or 'sample').}

\item{alpha}{Significance level for hypothesis testing.}

\item{tails}{Number of tails for hypothesis testing.}

\item{nu}{Hypothesized value to test the effect against.}

\item{to_return}{Type of output to return ('raw_output', 'print', or other).}

\item{model_object}{A model object used in the sensitivity analysis.}

\item{tested_variable}{Name of the variable being tested in the model.}
}
\value{
Depending on `to_return`,a list of analysis results or printed output
}
\description{
This function performs sensitivity analysis for non-linear models.
It is used in conjunction with `pkonfound()`, `konfound()`, and `mkonfound()`.
}
