% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkonfound.R
\name{tkonfound}
\alias{tkonfound}
\title{Perform tkonfound analysis}
\usage{
tkonfound(a, b, c, d, thr_p = 0.05, switch_trm = T)
}
\arguments{
\item{a}{cell a}

\item{b}{cell b}

\item{c}{cell c}

\item{d}{cell d}

\item{thr_p}{p-value threshhold}

\item{switch_trm}{whether to switch the treatment and control cases; defaults to FALSE; applicable only when non_linear - TRUE}
}
\value{
the output of tkonfound
}
\description{
Perform tkonfound analysis
}
\examples{
tkonfound(35, 17, 17, 38)
}
