% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_pse_auxiliary.R
\name{verify_reg_Gz}
\alias{verify_reg_Gz}
\title{Verify regression model with control variable Z}
\usage{
verify_reg_Gz(n_obs, sdx, sdy, sdz, rxy, rxz, rzy)
}
\arguments{
\item{n_obs}{number of observations}

\item{sdx}{standard deviation of X}

\item{sdy}{standard deviation of Y}

\item{sdz}{standard deviation of Z}

\item{rxy}{correlation coefficient between X and Y}

\item{rxz}{correlation coefficient between X and Z}

\item{rzy}{correlation coefficient between Z and Y}
}
\value{
list of model parameters
}
\description{
Verify regression model with control variable Z
}
