% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_pse_auxiliary.R
\name{verify_reg_Gzcv}
\alias{verify_reg_Gzcv}
\title{Verify regression model with control variable Z}
\usage{
verify_reg_Gzcv(n_obs, sdx, sdy, sdz, sdcv, rxy, rxz, rzy, rcvy, rcvx, rcvz)
}
\arguments{
\item{n_obs}{number of observations}

\item{sdx}{standard deviation of X}

\item{sdy}{standard deviation of Y}

\item{sdz}{standard deviation of Z}

\item{sdcv}{sd between C and V}

\item{rxy}{correlation coefficient between X and Y}

\item{rxz}{correlation coefficient between X and Z}

\item{rzy}{correlation coefficient between Z and Y}

\item{rcvy}{correlation coefficient between V and Y}

\item{rcvx}{correlation coefficient between V and X}

\item{rcvz}{correlation coefficient between V and Z}
}
\value{
list of model parameters
}
\description{
Verify regression model with control variable Z
}
