% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_pse_auxiliary.R
\name{verify_reg_uncond}
\alias{verify_reg_uncond}
\title{Verify unconditional regression model}
\usage{
verify_reg_uncond(n_obs, sdx, sdy, rxy)
}
\arguments{
\item{n_obs}{number of observations}

\item{sdx}{standard deviation of X}

\item{sdy}{standard deviation of Y}

\item{rxy}{correlation coefficient between X and Y}
}
\value{
list of model parameters
}
\description{
Verify unconditional regression model
}
