% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concord1.R
\docType{data}
\name{concord1}
\alias{concord1}
\title{Concord1 data}
\format{
A data.frame with 496 rows and 10 variables.
}
\description{
This data is from Hamilton (1983)
}
\references{
Hamilton, Lawrence C. 1983. Saving water: A causal model of household conservation. Sociological Perspectives 26(4):355-374.
}
