% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{konfound}
\alias{konfound}
\title{Perform sensitivity analysis on fitted models}
\usage{
konfound(model_object, tested_variable, alpha = 0.05, tails = 2)
}
\arguments{
\item{model_object}{output from a model (currently works for: lm)}

\item{tested_variable}{Variable associated with the unstandardized beta coefficient to be tested}

\item{alpha}{probability of rejecting the null hypothesis (defaults to 0.05)}

\item{tails}{integer whether hypothesis testing is one-tailed (1) or two-tailed (2; defaults to 2)}

\item{to_return}{whether to return the results in a data.frame (defaults to FALSE; instead results are printed to the console)}
}
\value{
prints the bias and the number of cases that would have to be replaced with cases for which there is no effect to invalidate the inference
}
\description{
For fitted models, this command calculates (1) how much bias there must be in an estimate to invalidate/sustain an inference; (2) the impact of an omitted variable necessary to invalidate/sustain an inference for a regression coefficient. Currently works for: models created with lm() (linear models).
}
\examples{
m1 <- lm(mpg ~ wt + hp, data = mtcars)
konfound(m1, wt)
}
