% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{launch_shiny}
\alias{launch_shiny}
\title{Open interactive web application for konfound}
\usage{
launch_shiny()
}
\value{
Launches a web browser
}
\description{
Open interactive web application for konfound
}
\details{
Open the Shiny interactive web application in a browser
}
