% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mkonfound}
\alias{mkonfound}
\title{Perform meta-analyses including sensitivity analysis}
\usage{
mkonfound(df, alpha = 0.05, tails = 2)
}
\arguments{
\item{df}{a data.frame with columns for the 'unstd_beta', 'std_err', 'n_obs', and 'n_covariates' arguments (see ?pkonfound for more information on these)}

\item{alpha}{probability of rejecting the null hypothesis (defaults to 0.05)}

\item{tails}{integer whether hypothesis testing is one-tailed (1) or two-tailed (2; defaults to 2)}
}
\value{
prints the bias and the number of cases that would have to be replaced with cases for which there is no effect to invalidate the inference for each of the cases in the data.frame
}
\description{
For fitted models, this command carries out sensitivity analysis for a number of models, when their parameters stored in a data.frame.
}
\examples{
df <- data.frame(unstd_beta = c(2, 10, 1.7), std_error = c(.3, 2.9, 1.5), n_obs = c(70, 405, 200), n_covs = c(3, 4, 1))
mkonfound(df)
}
