% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pkonfound}
\alias{pkonfound}
\title{Perform sensitivity analysis for published studies}
\usage{
pkonfound(unstd_beta, standard_error, n_obs, n_covariates = 3, alpha = 0.05,
  tails = 2)
}
\arguments{
\item{unstd_beta}{an unstandardized regression coefficient}

\item{standard_error}{the standard error of the estimate of the unstandardized regression coefficient}

\item{n_obs}{the number of observations in the sample}

\item{n_covariate}{the number of covariates in the regression model}
}
\value{
prints the bias and the number of cases that would have to be replaced with cases for which there is no effect to invalidate the inference
}
\description{
For published studies, this command calculates (1) how much bias there must be in an estimate to invalidate/sustain an inference; (2) the impact of an omitted variable necessary to invalidate/sustain an inference for a regression coefficient.
}
\examples{
pkonfound(unstd_beta = 2, standard_error = .4, n_obs = 100, n_covariates = 3)
pkonfound(unstd_beta = .4, standard_error = 2, n_obs = 100, n_covariates = 3)
}
