% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkonfound.R
\name{pkonfound}
\alias{pkonfound}
\title{Perform sensitivity analysis for published studies}
\usage{
pkonfound(unstd_beta, std_err, n_obs, n_covariates = 1, alpha = 0.05,
  tails = 2, nu = 0, to_return = "print",
  component_correlations = FALSE)
}
\arguments{
\item{unstd_beta}{an unstandardized regression coefficient}

\item{std_err}{the standard error of the estimate of the unstandardized regression coefficient}

\item{n_obs}{the number of observations in the sample}

\item{n_covariates}{the number of covariates in the regression model}

\item{alpha}{probability of rejecting the null hypothesis (defaults to 0.05)}

\item{tails}{integer whether hypothesis testing is one-tailed (1) or two-tailed (2; defaults to 2)}

\item{nu}{what hypothesis to be tested; defaults to testing whether unstd_beta is significantly different from 0}

\item{to_return}{whether to return a data.frame (by specifying this argument to euqal "df") or a plot ("plot"); default is to print the output to the console}

\item{component_correlations}{whether to return the component correlations as part of the correlation-based approach}
}
\value{
prints the bias and the number of cases that would have to be replaced with cases for which there is no effect to invalidate the inference
}
\description{
For published studies, this command calculates (1) how much bias there must be in an estimate to invalidate/sustain an inference; (2) the impact of an omitted variable necessary to invalidate/sustain an inference for a regression coefficient.
}
\examples{
pkonfound(2, .4, 100, 3)
pkonfound(.4, 2, 100, 3)
}
